requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
vllm>=0.4.0
deepspeed>=0.14.0
accelerate>=0.30.0
sympy>=1.12
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.24.0
jsonlines>=4.0.0
anthropic>=0.25.0
